refactor : update service and repository methods for consistency and clarity; enhance user and pet deletion logic

+++ src/repositories/pet/index.js
@@ -5,7 +5,7 @@ export const getPetById = async (id) => {
   return pet;
 };
 
-export const deletePetById = async (id) => {
-  const pet = await Pet.findByIdAndDelete(id);
+export const deletePetById = async (petId, userId) => {
+  const pet = await Pet.findByIdAndDelete({ id: petId, owner: userId });
   return pet;
 };
